Module google_videointelligence1_beta1::api

Structs

CloudVideoIntelligence
Central instance to access all CloudVideoIntelligence related resource activities

GoogleCloudVideointelligenceV1beta1_AnnotateVideoRequest
Video annotation request.

GoogleCloudVideointelligenceV1beta1_VideoContext
Video context and/or feature-specific parameters.

GoogleCloudVideointelligenceV1beta1_VideoSegment
Video segment.

GoogleLongrunning_Operation
This resource represents a long-running operation that is the result of a network API call.

GoogleRpc_Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. The error model is designed to be:

VideoAnnotateCall
Performs asynchronous video annotation. Progress and results can be retrieved through the google.longrunning.Operations interface. Operation.metadata contains AnnotateVideoProgress (progress). Operation.response contains AnnotateVideoResponse (results).

VideoMethods
A builder providing access to all methods supported on video resources. It is not used directly, but through the CloudVideoIntelligence hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.